Move most of data sources and repositories to a separated package: source
