+++ app/src/main/java/com/yurii/youtubemusic/screens/authorization/AuthenticationFragment.kt
@@ -7,8 +7,8 @@ import android.widget.Toast
 import com.google.android.gms.common.api.ApiException
 import com.yurii.youtubemusic.R
 import com.yurii.youtubemusic.databinding.FragmentAuthenticationBinding
-import com.yurii.youtubemusic.utilities.DoesNotHaveRequiredScopes
-import com.yurii.youtubemusic.utilities.GoogleAccount
+import com.yurii.youtubemusic.source.DoesNotHaveRequiredScopes
+import com.yurii.youtubemusic.source.GoogleAccount
 import com.yurii.youtubemusic.utilities.TabFragment
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
