+++ app/src/main/java/com/yurii/youtubemusic/services/downloader/MusicDownloaderService.kt
@@ -4,10 +4,9 @@ import android.app.Service
 import android.content.Intent
 import android.os.Binder
 import android.os.IBinder
-import android.util.Log
 import com.yurii.youtubemusic.models.*
 import com.yurii.youtubemusic.services.media.MediaLibraryManager
-import com.yurii.youtubemusic.utilities.MediaManagerDomain
+import com.yurii.youtubemusic.source.MediaManagerDomain
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
