+++ app/src/main/java/com/yurii/youtubemusic/source/EqualizerPreferences.kt
@@ -1,4 +1,4 @@
-package com.yurii.youtubemusic.utilities
+package com.yurii.youtubemusic.source
 
 import android.content.Context
 import com.google.gson.Gson
