+++ app/src/main/java/com/yurii/youtubemusic/source/MediaRepository.kt
@@ -1,4 +1,4 @@
-package com.yurii.youtubemusic.utilities
+package com.yurii.youtubemusic.source
 
 import com.yurii.youtubemusic.db.*
 import com.yurii.youtubemusic.models.MediaItem
